updated nvim callbacks
textDocument/references callback now falls back to selecting the closest
item in the file, if the current line was not found on the qf list.
extended that functionality to other LSP callbacks.

also changed the location of zsh git-prompt plugin. no idea how to set a
git submodule in a bare repo though, so installing git-promt.sh has to
be done manually.